added graph and a rocket fall

+++ src/Point2D.py
@@ -1,5 +1,5 @@
 class Point2D:
-    def __init__(self, x: int, y: int) -> None:
+    def __init__(self, x: float, y: float) -> None:
         self.x = x
         self.y = y
